Refactor prometheus labelers to be more composable
